pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contract-generator"
version = "0.1.0"
description = "Generate modern fillable contract PDFs from YAML templates."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "jinja2>=3.1",
  "markdown-it-py>=3.0",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "reportlab>=4.0",
  "pypdf>=5.0",
]

[project.scripts]
contract-generator = "contract_generator.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["contract_generator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
